# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Movie APP",
    "description": "Open API for movie app"
  },
  "paths": {
    "/add_new_movie": {
      "post": {
        "tags": [
          "Add new movie"
        ],
        "description": "Add new movie in database.",
        "application/x-www-form-urlencoded":{
          "schema": {
            "type": "object",
            "properties": {
              "movie_name": {"type": "string"},
              "locations_and_timings": {"type": "object"}
              },
            "required": true

          }
        },
        "responses": {
          "200": {
            "description": "It will add the movie name and informations in database.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "example": {"data": [], "status": "success"}
                }
              }
            }
          }
        }
      }
    },
    "/get_all_movies": {
      "get": {
        "tags": [
          "Get all movies"
        ],
        "description": "Returns hello in specified language",
        "parameters": [
          {
            "in": "query",
            "name": "movie_name",
            "description": "Filter by movie name.",
            "schema": {
              "type": "string",
              "example": "abcd"
            }
          },
          {
            "in": "query",
            "name": "timing",
            "description": "Filter by timing.",
            "schema": {
              "type": "string",
              "example": "13:00"
            }
          },
          {
            "in": "query",
            "name": "location",
            "description": "Filter by location",
            "schema": {
              "type": "string",
              "example": "delhi"
            }
          }

        ],
        "responses": {
          "200": {
            "description": "It will return a list of dict according to your filters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "example": {"data": [], "status": "success"}
                }
              }
            }
          }
        }
      }
    },
    "/delete_movie/{movie_name}": {
      "delete": {
        "tags": [
          "Delete Movie by name"
        ],
        "description": "Returns hello in specified language",
        "parameters": [
          {
            "in": "path",
            "name": "movie_name",
            "description": "Filter by movie name.",
            "schema": {
              "type": "string",
              "example": "abcd"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "It will return a list of dict according to your filters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "example": {"data": [], "status": "success"}
                }
              }
            }
          }
        }
      }
    }
  }
}